# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-Migration Assist API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreateLeadInput {
  fullName: string;
  email: string;
  whatsapp?: string;
  nationality: string;
  countryOfResidence?: string;
  currentlyInSouthAfrica?: boolean;
  /** valid | expired | lost | none */
  passportStatus?: string;
  visaHistory?: string;
  /** valid | expired | overstay | undesirable | prohibited | unknown */
  immigrationSituation: string;
  visaExpiryDate?: string;
  exitDate?: string;
  borderDocumentIssued?: string;
  /** medical | accident | family_emergency | admin_delay | other */
  overstayReason?: string;
  /** yes | some | no */
  hasSupportingDocuments?: string;
  previousOverstay?: string;
  /** email | whatsapp | phone */
  preferredContactMethod?: string;
  consentAccepted: boolean;
}

export interface UpdateLeadInput {
  /** NEW | REVIEWED | NEEDS_FOLLOW_UP | WAITLISTED | NOT_RELEVANT */
  leadStatus?: string;
  adminNotes?: string | null;
}

export interface Lead {
  id: string;
  referenceNumber: string;
  fullName?: string | null;
  email?: string | null;
  whatsapp?: string | null;
  nationality?: string | null;
  countryOfResidence?: string | null;
  currentlyInSouthAfrica?: boolean | null;
  passportStatus?: string | null;
  visaHistory?: string | null;
  immigrationSituation?: string | null;
  visaExpiryDate?: string | null;
  exitDate?: string | null;
  borderDocumentIssued?: string | null;
  overstayReason?: string | null;
  hasSupportingDocuments?: string | null;
  previousOverstay?: string | null;
  internalClassification?: string | null;
  leadScore?: number | null;
  leadCategory?: string | null;
  /** HIGH_PRIORITY | MEDIUM_PRIORITY | LOW_PRIORITY */
  leadPriority?: string | null;
  /** NEW | REVIEWED | NEEDS_FOLLOW_UP | WAITLISTED | NOT_RELEVANT */
  leadStatus: string;
  adminNotes?: string | null;
  preferredContactMethod?: string | null;
  consentAccepted: boolean;
  consentTimestamp?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface StatsSummary {
  totalAssessments: number;
  last24Hours: number;
  avgScore: number;
  byCategory: CategoryCount[];
  byPriority: CategoryCount[];
  byStatus: CategoryCount[];
}

export type AnalyticsEventInputPayload = { [key: string]: unknown } | null;

export interface AnalyticsEventInput {
  /** assessment_started | assessment_completed | classification_result | document_upload */
  eventName: string;
  referenceNumber?: string | null;
  payload?: AnalyticsEventInputPayload;
}

export interface AnalyticsEventResponse {
  id: string;
  eventName: string;
  createdAt: string;
}

export type ListLeadsParams = {
  /**
   * @minimum 1
   * @maximum 500
   */
  limit?: number;
  /**
   * Filter by HIGH_PRIORITY | MEDIUM_PRIORITY | LOW_PRIORITY
   */
  priority?: string;
  /**
   * Filter by NEW | REVIEWED | NEEDS_FOLLOW_UP | WAITLISTED | NOT_RELEVANT
   */
  status?: string;
  nationality?: string;
  /**
   * Filter by immigrationSituation enum value
   */
  situation?: string;
};