# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: E-Migration Assist API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: leads
    description: Pre-launch immigration assessment leads
  - name: stats
    description: Aggregate stats
  - name: analytics
    description: Funnel analytics events
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /leads:
    post:
      operationId: createLead
      tags: [leads]
      summary: Submit a completed immigration assessment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadInput"
      responses:
        "201":
          description: Lead created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "200":
          description: Existing lead updated (duplicate email or whatsapp detected)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    get:
      operationId: listLeads
      tags: [leads]
      summary: List recent leads (admin overview) with optional filters
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
        - name: priority
          in: query
          required: false
          description: Filter by HIGH_PRIORITY | MEDIUM_PRIORITY | LOW_PRIORITY
          schema:
            type: string
        - name: status
          in: query
          required: false
          description: Filter by NEW | REVIEWED | NEEDS_FOLLOW_UP | WAITLISTED | NOT_RELEVANT
          schema:
            type: string
        - name: nationality
          in: query
          required: false
          schema:
            type: string
        - name: situation
          in: query
          required: false
          description: Filter by immigrationSituation enum value
          schema:
            type: string
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
  /leads/{referenceNumber}:
    get:
      operationId: getLeadByReference
      tags: [leads]
      summary: Look up a lead by reference number
      parameters:
        - name: referenceNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lead found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
  /leads/by-id/{id}:
    get:
      operationId: getLeadById
      tags: [leads]
      summary: Look up a lead by internal UUID (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lead found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update admin-only fields (status, notes)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadInput"
      responses:
        "200":
          description: Lead updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
  /stats/summary:
    get:
      operationId: getStatsSummary
      tags: [stats]
      summary: Aggregate stats (totals, priority breakdown)
      responses:
        "200":
          description: Stats summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsSummary"
  /analytics/events:
    post:
      operationId: trackAnalyticsEvent
      tags: [analytics]
      summary: Record a funnel analytics event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyticsEventInput"
      responses:
        "201":
          description: Event recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsEventResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    CreateLeadInput:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        whatsapp:
          type: string
        nationality:
          type: string
        countryOfResidence:
          type: string
        currentlyInSouthAfrica:
          type: boolean
        passportStatus:
          type: string
          description: 'valid | expired | lost | none'
        visaHistory:
          type: string
        immigrationSituation:
          type: string
          description: 'valid | expired | overstay | undesirable | prohibited | unknown'
        visaExpiryDate:
          type: string
          format: date
        exitDate:
          type: string
          format: date
        borderDocumentIssued:
          type: string
        overstayReason:
          type: string
          description: 'medical | accident | family_emergency | admin_delay | other'
        hasSupportingDocuments:
          type: string
          description: 'yes | some | no'
        previousOverstay:
          type: string
        preferredContactMethod:
          type: string
          description: 'email | whatsapp | phone'
        consentAccepted:
          type: boolean
      required:
        - fullName
        - email
        - nationality
        - immigrationSituation
        - consentAccepted
    UpdateLeadInput:
      type: object
      properties:
        leadStatus:
          type: string
          description: 'NEW | REVIEWED | NEEDS_FOLLOW_UP | WAITLISTED | NOT_RELEVANT'
        adminNotes:
          type: string
          nullable: true
    Lead:
      type: object
      properties:
        id:
          type: string
        referenceNumber:
          type: string
        fullName:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        whatsapp:
          type: string
          nullable: true
        nationality:
          type: string
          nullable: true
        countryOfResidence:
          type: string
          nullable: true
        currentlyInSouthAfrica:
          type: boolean
          nullable: true
        passportStatus:
          type: string
          nullable: true
        visaHistory:
          type: string
          nullable: true
        immigrationSituation:
          type: string
          nullable: true
        visaExpiryDate:
          type: string
          nullable: true
        exitDate:
          type: string
          nullable: true
        borderDocumentIssued:
          type: string
          nullable: true
        overstayReason:
          type: string
          nullable: true
        hasSupportingDocuments:
          type: string
          nullable: true
        previousOverstay:
          type: string
          nullable: true
        internalClassification:
          type: string
          nullable: true
        leadScore:
          type: integer
          nullable: true
        leadCategory:
          type: string
          nullable: true
        leadPriority:
          type: string
          nullable: true
          description: 'HIGH_PRIORITY | MEDIUM_PRIORITY | LOW_PRIORITY'
        leadStatus:
          type: string
          description: 'NEW | REVIEWED | NEEDS_FOLLOW_UP | WAITLISTED | NOT_RELEVANT'
        adminNotes:
          type: string
          nullable: true
        preferredContactMethod:
          type: string
          nullable: true
        consentAccepted:
          type: boolean
        consentTimestamp:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - referenceNumber
        - consentAccepted
        - leadStatus
        - createdAt
        - updatedAt
    CategoryCount:
      type: object
      properties:
        category:
          type: string
        count:
          type: integer
      required: [category, count]
    StatsSummary:
      type: object
      properties:
        totalAssessments:
          type: integer
        last24Hours:
          type: integer
        avgScore:
          type: number
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCount"
        byPriority:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCount"
        byStatus:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCount"
      required: [totalAssessments, last24Hours, avgScore, byCategory, byPriority, byStatus]
    AnalyticsEventInput:
      type: object
      properties:
        eventName:
          type: string
          description: 'assessment_started | assessment_completed | classification_result | document_upload'
        referenceNumber:
          type: string
          nullable: true
        payload:
          type: object
          nullable: true
          additionalProperties: true
      required: [eventName]
    AnalyticsEventResponse:
      type: object
      properties:
        id:
          type: string
        eventName:
          type: string
        createdAt:
          type: string
      required: [id, eventName, createdAt]