# lib/api-zod/src/generated/types/analyticsEventInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-Migration Assist API
 * OpenAPI spec version: 0.2.0
 */
import type { AnalyticsEventInputPayload } from "./analyticsEventInputPayload";

export interface AnalyticsEventInput {
  /** assessment_started | assessment_completed | classification_result | document_upload */
  eventName: string;
  referenceNumber?: string | null;
  payload?: AnalyticsEventInputPayload;
}