# lib/api-zod/src/generated/types/categoryCount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-Migration Assist API
 * OpenAPI spec version: 0.2.0
 */

export interface CategoryCount {
  category: string;
  count: number;
}