# lib/api-zod/src/generated/types/createLeadInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-Migration Assist API
 * OpenAPI spec version: 0.2.0
 */

export interface CreateLeadInput {
  fullName: string;
  email: string;
  whatsapp?: string;
  nationality: string;
  countryOfResidence?: string;
  currentlyInSouthAfrica?: boolean;
  /** valid | expired | lost | none */
  passportStatus?: string;
  visaHistory?: string;
  /** valid | expired | overstay | undesirable | prohibited | unknown */
  immigrationSituation: string;
  visaExpiryDate?: Date;
  exitDate?: Date;
  borderDocumentIssued?: string;
  /** medical | accident | family_emergency | admin_delay | other */
  overstayReason?: string;
  /** yes | some | no */
  hasSupportingDocuments?: string;
  previousOverstay?: string;
  /** email | whatsapp | phone */
  preferredContactMethod?: string;
  consentAccepted: boolean;
}