# lib/api-zod/src/generated/types/lead.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-Migration Assist API
 * OpenAPI spec version: 0.2.0
 */

export interface Lead {
  id: string;
  referenceNumber: string;
  fullName?: string | null;
  email?: string | null;
  whatsapp?: string | null;
  nationality?: string | null;
  countryOfResidence?: string | null;
  currentlyInSouthAfrica?: boolean | null;
  passportStatus?: string | null;
  visaHistory?: string | null;
  immigrationSituation?: string | null;
  visaExpiryDate?: string | null;
  exitDate?: string | null;
  borderDocumentIssued?: string | null;
  overstayReason?: string | null;
  hasSupportingDocuments?: string | null;
  previousOverstay?: string | null;
  internalClassification?: string | null;
  leadScore?: number | null;
  leadCategory?: string | null;
  /** HIGH_PRIORITY | MEDIUM_PRIORITY | LOW_PRIORITY */
  leadPriority?: string | null;
  /** NEW | REVIEWED | NEEDS_FOLLOW_UP | WAITLISTED | NOT_RELEVANT */
  leadStatus: string;
  adminNotes?: string | null;
  preferredContactMethod?: string | null;
  consentAccepted: boolean;
  consentTimestamp?: string | null;
  createdAt: string;
  updatedAt: string;
}