# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-Migration Assist API
 * OpenAPI spec version: 0.2.0
 */
import type { CategoryCount } from "./categoryCount";

export interface StatsSummary {
  totalAssessments: number;
  last24Hours: number;
  avgScore: number;
  byCategory: CategoryCount[];
  byPriority: CategoryCount[];
  byStatus: CategoryCount[];
}