# lib/api-zod/src/generated/types/updateLeadInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-Migration Assist API
 * OpenAPI spec version: 0.2.0
 */

export interface UpdateLeadInput {
  /** NEW | REVIEWED | NEEDS_FOLLOW_UP | WAITLISTED | NOT_RELEVANT */
  leadStatus?: string;
  adminNotes?: string | null;
}